Initial header commit
Create the initial header. This isn't done yet: I see the formatting I was hoping for on the body won't work as I initially expected in terms of the background color, but that should be a fairly easy fix later since I plan to put everything in containers. The formatting on <a> also needs worked out.

index.html
- Create link to stylesheet

stylesheet.css
- Set <body> to be a flex box with some general characteristics
- Create initial classes for the header and fill in some basic information for them
- <a> needs some more work done on it