Initial introduction/hero html code
Just the initial formatting for the introduction (or as the Odin project calls the area, the "hero'). The <img> is not showing how I want it to, but I'm not to stress about getting it perfect right now.

A note about my writing style. In documentation I like to use a / between things to act as an and so I don't have to rewrite things multiple times.
- If I say something like .object-left/right it stands for
 - .object-left
 - .object-right
- I'm noting this here because I do it in this documentation, as you will see later

index.html
- Add in some specific classes for formatting purposes to the introduction section
 - Main and secondary text
 - Button
- Tried some basic image formatting, but so far it hasn't done much

stylesheet.css
- Added a bunch of classes for the various areas of the introduction and fill them with formatting
 - .introduction-left/right (filled out, they already existed)
 - .hero-main-text
 - .hero-secondary-text
 - .hero-sign-up-button
- Also a couple of misc changes to the header's .link-list and getting rid of the default margin on the body (I think this is due to it being a flex-box)